Crate rdrand

An implementation of random number generator based on rdrand instruction.

rdrand is claimed to be a cryptographically secure PRNG. It is much faster than OsRng (and slower than StdRng), but is only supported on more recent Intel processors.

The generator provided by this crate is a viable replacement to any std::rand generator, however, since nobody has audited Intel hardware yet, the usual disclaimers apply.

Structs

RdRand

Enums

Error