Struct rdrand::RdRand

pub struct RdRand(_);

A cryptographically secure pseudo-random number generator.

This generator is a viable replacement to any std::rand generator, however, since nobody has audited Intel hardware yet, the usual disclaimers apply.

It is much faster than OsRng (and slower than StdRng), but is only supported on more recent (since Ivy Bridge) Intel processors.

Methods

impl RdRand

fn new() -> Result<RdRand, Error>

Build a generator object. The function will only succeed if rdrand instruction can be successfully used.

fn next_u16(&self) -> u16

Generate a u16 value.